fix: re-orged print commands
