+++ src/test/e2e/cmd/testdata/dev/print/resources.golden
