+++ src/test/e2e/cmd/testdata/dev/print/validation.golden
